Refactor into shared support module.
